gh-138013: Move buffered tests to test_bufferedio
This moves test support code to `test_io._support` and buffered test
cases to `test_io.test_bufferedio` via copy/paste code movement then
adjusts imports as needed (remove unneded + add all required).
